Feat: Implement rich console output.

+++ src/bpydevutil/install.py
@@ -2,8 +2,20 @@
 
 import shutil
 from pathlib import Path
-
 from typing import Optional
+
+import typer
+from rich import print
+from rich.panel import Panel
+from rich.progress import track
+
+
+def render_settings(addons_src, addons_install_dir, excluded_addons):
+    src_string = f"Addon Sources = {addons_src}"
+    addons_install_string = f"Addon Install Directory = {addons_install_dir}"
+    excluded_addons_string = f"Excluded Addons = {excluded_addons}"
+
+    return "\n".join([src_string, addons_install_string, excluded_addons_string])
 
 
 class InstallAddonsFromSource:
@@ -22,6 +34,14 @@ class InstallAddonsFromSource:
         self.addons_src = addons_src
         self.addons_install_dir = addons_install_dir
         self.excluded_addons = excluded_addons
+
+        print(
+            Panel.fit(
+                render_settings(addons_src, addons_install_dir, excluded_addons),
+                title="[orange3]Installer Settings[/orange3]",
+                border_style="yellow",
+            )
+        )
 
     def get_addon_srcs(self) -> list[Path]:
         """Get a list of addon source paths that need to be installed.
@@ -49,7 +69,7 @@ class InstallAddonsFromSource:
                 if "bl_info" in path.read_text():
                     return True
 
-            return False
+            print(f"[italic]Skipping <{path.name}>, no [yellow]bl_info[/yellow] found.[/italic]")
 
         if self.excluded_addons:
             addon_srcs = [
@@ -57,6 +77,10 @@ class InstallAddonsFromSource:
             ]
         else:
             addon_srcs = [path for path in self.addons_src.iterdir() if is_python(path)]
+
+        if not addon_srcs:
+            print(f"[red]There are no addon sources inside the directory [{self.addons_src}][/red]")
+            raise typer.Abort()
 
         return addon_srcs
 
@@ -73,11 +97,15 @@ class InstallAddonsFromSource:
             path for path in self.addons_install_dir.iterdir() if path.exists() and path.name in addon_names
         ]
 
-        for path in delete_paths:
-            if path.is_file():
-                path.unlink()
-            else:
-                shutil.rmtree(path)
+        if not delete_paths:
+            return
+
+        for _ in track(delete_paths, show_speed=False, description="Removing old versions..."):
+            for path in delete_paths:
+                if path.is_file():
+                    path.unlink()
+                else:
+                    shutil.rmtree(path)
 
     def install_addons(self, addons: list[Path]) -> None:
         """Install addon sources to the addon directory.
@@ -87,8 +115,11 @@ class InstallAddonsFromSource:
 
         """
 
-        for path in addons:
-            if path.is_file():
-                shutil.copyfile(path, Path(self.addons_install_dir / path.name))
-            else:
-                shutil.copytree(path, Path(self.addons_install_dir / path.name))
+        for _ in track(addons, description="Installing addons..."):
+            for path in addons:
+                if path.is_file():
+                    shutil.copyfile(path, Path(self.addons_install_dir / path.name))
+                else:
+                    shutil.copytree(path, Path(self.addons_install_dir / path.name))
+
+        print("[green]Done![/green]")
